LayoutTests/platform/ios/accessibility/display-contents/table-section-elements-expected.txt

This test ensures the acessibility tree is correct when table section elements (e.g. thead, tbody) have display:contents.


{StaticText AXLabel: Author}

{StaticText AXLabel: Title}

{StaticText AXLabel: Stephen Hawking}

{StaticText AXLabel: A Brief History of Time}

{StaticText AXLabel: Carl Sagan}

{StaticText AXLabel: Cosmos}

PASS successfullyParsed is true

TEST COMPLETE
Author	Title
Stephen Hawking	A Brief History of Time
Carl Sagan	Cosmos
